+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Persian (Farsi) translation [#72](https://github.com/rlanvin/php-rrule/pull/72)
 - Support for Windows timezone [#69](https://github.com/rlanvin/php-rrule/issues/69)
 
+### Fixed
+
+- Fixed RRule::createFromRfcString not detecting RSet properly if the rule was lowercase
+
 ## [2.1.0] - 2019-05-30
 
 ### Fixed